增加@CacheDelete, 增加了使用AOP进行删除缓存的实现。

+++ src/main/java/com/jarvis/cache/memcache/CachePointCut.java
@@ -38,6 +38,9 @@ public class CachePointCut extends AbstractCacheManager<Serializable> {
      */
     @Override
     public void delete(String key) {
+        if(null == memcachedClient || null == key) {
+            return;
+        }
         try {
             memcachedClient.delete(key);
             this.getAutoLoadHandler().resetAutoLoadLastLoadTime(key);
